+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Master
 
+ - CLI: Add support for reading BGEO files
  - Lib: Allow passing in an existing `SurfaceReconstruction` to reuse allocated memory (currently only memory for the resulting surface mesh is reused) (https://github.com/w1th0utnam3/splashsurf/pull/7)
  - Lib: Add `Default` trait bound to `Index` and `Real` types